tasklists_auth

+++ app/Http/Controllers/TasklistsController.php
@@ -19,9 +19,25 @@ class TasklistsController extends Controller
      // getでtasklists/にアクセスされた場合の『一覧表示処理』
     public function index()
     {
-        $tasklists = Tasklist::all();
+        /* 以下はすべてのユーザーのタスクをviewに渡す処理
+        // $tasklists = Tasklist::all();
 
-        return view('tasklists.index', ['tasklists' => $tasklists, ]);
+        // return view('tasklists.index', ['tasklists' => $tasklists, ]);
+        */
+        
+        // 以下は個々のユーザーだけのタスクをviewに渡す処理
+        $data = [];
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasklists = $user->tasklists()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasklists' => $tasklists,
+            ];
+        }
+
+        return view('tasklists.index', $data);
     }
 
     /**
@@ -32,9 +48,20 @@ class TasklistsController extends Controller
      // getでtasklists/createにアクセスされた場合の『新規登録画面表示処理』
     public function create()
     {
-        $tasklist = new Tasklist;
-
-        return view('tasklists.create', ['tasklist' => $tasklist, ]);
+        if(\Auth::check()){
+            //$tasklist = \App\Tasklist::find($id);
+            //if (\Auth::user()->id === $tasklist->user_id) {
+            // if ($request->user()->id === $tasklist->user_id) {
+                //return view('tasklists.create', ['tasklist' => $tasklist,]);
+                return view('tasklists.create');
+            //} else {
+            //    return redirect('/');
+           // }
+        //} else {
+        //    return redirect('/');
+        } else {
+            return view('/');
+        }
     }
 
     /**
@@ -52,11 +79,15 @@ class TasklistsController extends Controller
             'content' => 'required | max:255',
         ]);
 
-        $tasklist = new Tasklist;
-        $tasklist->status = $request->status;  // 追加
-        $tasklist->content = $request->content;
-        $tasklist->save();
-
+      if(\Auth::check()){
+            $user = \Auth::user();
+    
+            $tasklist = new Tasklist;
+            $tasklist->status = $request->status;  // 追加
+            $tasklist->content = $request->content;
+            $tasklist->user_id = $user->id; // 外部キー制約
+            $tasklist->save();
+      }
         return redirect('/'); 
     }
 
@@ -69,8 +100,25 @@ class TasklistsController extends Controller
      // getでtasklists/idにアクセスされた場合の『取得表示処理』
     public function show($id)
     {
-        $tasklist = Tasklist::find($id);
-        return view('tasklists.show', ['tasklist' => $tasklist]);
+        // 以下はすべてのユーザーのタスクをviewに渡す処理
+        // $tasklist = Tasklist::find($id);
+        // return view('tasklists.show', ['tasklist' => $tasklist]);
+
+
+        // 以下は個々のユーザーだけのタスクをviewに渡す処理
+        $data = [];
+        if(\Auth::check()){
+            $user = \Auth::user();
+            $tasklist = $user->tasklists()->where('id',$id)->get();
+            //$tasklist = $user->tasklists()->where('id',$id)->get();
+//dd($tasklist[0]->id);
+            $data = [
+                'user' => $user,
+                'tasklist' => $tasklist[0],
+            ];
+        }
+
+        return view('tasklists.show', $data);
     }
 
     /**
@@ -82,8 +130,16 @@ class TasklistsController extends Controller
      // getでtasklists/id/editにアクセスされた場合の『更新画面表示処理』
     public function edit($id)
     {
-        $tasklist = Tasklist::find($id);
-        return view('tasklists.edit', ['tasklist' => $tasklist,]);
+         if(\Auth::check()){
+            $tasklist = Tasklist::find($id);
+            if ($request->user()->id === $tasklist->user_id) {
+                return view('tasklists.edit', ['tasklist' => $tasklist,]);
+            } else {
+                return redirect('/');
+            }
+         } else {
+             return redirect('/');
+         }
     }
 
     /**
@@ -102,12 +158,20 @@ class TasklistsController extends Controller
             'content' => 'required | max:255',
         ]);
 
-        $tasklist = Tasklist::find($id);
-        $tasklist->status = $request->status;  // 追加
-        $tasklist->content = $request->content;
-        $tasklist->save();
+        if(\Auth::check()){
+            $tasklist = Tasklist::find($id);
 
+            if ($request->user()->id === $tasklist->user_id) {
+                $tasklist->status = $request->status;  // 追加
+                $tasklist->content = $request->content;
+                $tasklist->user_id = $request->user()->id; // 外部キー制約
+                $tasklist->save();
+            }
+        }
+        
         return redirect('/');
+
+
     }
 
     /**
@@ -119,8 +183,14 @@ class TasklistsController extends Controller
      // deleteでtasklists/idにアクセスされた場合の『削除処理』
     public function destroy($id)
     {
-        $tasklist = Tasklist::find($id);
-        $tasklist->delete();
+        //$tasklist = Tasklist::find($id);
+        //$tasklist->delete();
+
+        $tasklist = \App\Tasklist::find($id);
+
+        if (\Auth::user()->id === $tasklist->user_id) {
+            $tasklist->delete();
+        }
 
         return redirect('/');
     }
